requirements.txt
datasets
diffusers==0.24.0
torch==2.1.1
torchvision==0.16.1
huggingface-hub
numpy
omegaconf
opencv-contrib-python
scikit-learn
transformers
tqdm
hydra-core